germinal command line parsing
Hi Jason

Thanks for providing the germinal server. While trying it, I noticed, that the
command line option for the configuration file does not work and the short
options collides with the one for the cert. Also the help option does not work
when the default configuration file is not present.

Below is how i got it to work. Also the "i" in GERMINAL_CONFiG should be
uppercase.

thanks
Tobias

 server.lisp | 19 ++++++++++---------
 1 file changed, 10 insertions(+), 9 deletions(-)