Add middleware to log requests and validate server name
A middleware is a function that takes a handler and returns another handler with some
added behavior. A handler is a function that takes a germinal:request and returns a
germinal:response. Middleware can be added by modifying *germinal-middleware*.

This middleware implementation is modeled after clack / ring (clojure).

Two middleware are provided:

- The access-log middleware logs requests to *standard-output* using a simple log format
- The validate-server-name middleware ensures that the requested host
  matches *germinal-server-name*. This supports deployment on servers that serve
  multiple domains.

The access-log middleware is turned on by default.