Fix path url encoding
Previously, the entire path was encoded with quri:url-encode, which encodes / as
%2F. This meant that an absolute path such as "/one/two and/three.gmi" would be rendered as
"/%2Fone%2Ftwo%20and%2Fthree.gmi". This works fine because of the corresponding call
to quri:url-decode, but it means clients end up requesting pretty weird URLs.

While quri:url-encode is correct according to the RFC1866, most server
implementations I've seen don't encode slashes. I believe the spirit of the gemini
spec is to behave like encodeURIComponent in javascript: ensure that spaces and
other reserved characters besides slashes are urlencoded.

This commit implements that behavior: path components are url-encoded, but slashes
are kept as-is. This commit also makes path construction work correctly on
e.g. windows by ensuring that we use "/" for URLs rather than relying on the
platform-specific output of namestring (untested).